Continued rewrite of metro automated setup -- now called "setup".
My goal is to make this script much simpler, but also much more robust.
It will try to do less, but do it reliably, rather than trying to do
too much (I think the previous script was too ambitious, unmaintained
and fragile.)

This is a work in progress. This setup script doesn't fully set up
a build yet -- but it's getting close.